Move buffers to separate project before extraction
