+++ src/Connectors.Batching/InMemory/Buffer.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using System.Threading;
 
-namespace CluedIn.Connector.AzureDataLake
+namespace CluedIn.Connectors.Batching.InMemory
 {
     internal class Buffer<T> : IDisposable
     {
